Refactor: separate structs definitions to driver/entities.go
